Add files via upload
I recently completed an exciting project on DataCamp titled "Investigating Netflix Movies and Guest Stars in The Office." This project provided a unique opportunity to explore and analyze the intersection of two pop culture phenomena: the vast library of Netflix movies and the beloved TV series "The Office."

The primary objective of this project was to gain insights into the presence of "The Office" cast members as guest stars in Netflix movies. As a cult classic sitcom, "The Office" boasts a memorable ensemble cast, and many of its actors have gone on to work in various film projects. Therefore, I set out to uncover which members of the Dunder Mifflin team had made appearances in Netflix movies.

The project journey began with data collection. I leveraged my data analysis skills to gather relevant information on "The Office" cast members and their filmography, particularly within the Netflix catalog. This involved data extraction, cleaning, and organization to create a structured dataset for further analysis.

The analysis phase was perhaps the most exciting part of this project. I used data manipulation and visualization techniques to identify patterns and trends. Through Python programming and libraries such as pandas and Matplotlib, I crafted insightful visualizations that helped me understand the relationships between "The Office" actors and their roles in Netflix movies.